MINOR: Update log statements in alterBrokerConfigs/alterTopicConfigs methods
